v2.1 (#159)
* Key Memory-Leak Prevention

Deletes the key and cipher after it's been used completely from memory. This prevents searching through memory for the key itself.

* Update doggo_ransomware.py

* Error fix

The line of code in question was prone to error if the computer did not have a battery associated with it. 

psutil.sensors_battery() can return a NoneType object.

* Nonetype return fixes

List index out of range is fixed.

* Remove unused libraries

sys, subprocess, and shutil were unused at this time.

May be used for future use, but, was removed.

* Remove unused libraries; antivm.py

sys was not used.

* Remove unused libraries; rose_rat.py

Removed a "Thread" object from the threading library, and eventlet

* Remove unused libraries; rose_builder.pyw

Removed time, platform, and the pyqtSlot ("Slot") reference from the PyQt5 library.

* Rename to webhook.py (__webhook.py)

Running it may cause errors there.

* Update for the __webhook.py rename

* Update for the __webhook.py rename

* Update for the __webhook.py rename

* Update for the __webhook.py rename

* Update for the __webhook.py rename

* Update for the __webhook.py rename

* Remove unused reference. (knight_rat.py)

pygame was unused, pygame.camera stays.

* Use wmi (sysinf.py)

This part of the code was the most prone to errors, as cpuinfo hanged at some points.

WMI is a little bit faster, and doesn't hang.

* Make the screen var in the embed look better.

* Remove py-cpuinfo, add wmi

WMI is now used for CPU Information, py-cpuinfo hung.

* REM statement changed to fit the python version actually installed.

* .

* sysinfo.py little update.

* v2.1 README

* v2.1